requirements.txt
aiosignal==1.3.1
altair==5.1.2
ansi2html==1.8.0
async==0.6.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.2
distlib==0.3.6
extra-streamlit-components==0.1.60
filelock==3.12.2
Flask==2.2.5
Flask-Login==0.6.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
platformdirs==3.8.0
plotly==5.17.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyarrow==14.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.22
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
validators==0.22.0
virtualenv==20.23.1
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
